Don't send entity list to client.
Instead of sending the entity list, the server sends Lua objects that describe the entities. Both the client and server
prepare the same the "map" data, then the client creates entities. The server can, but does not currently, create
entities. This brings the network traffic down by a very large amount.