Start working on instancing optimization.
Adds a new render object type for instanced objects and some code to
find objects that are instanced. The immediate effect of this is that
instanced objects are rendered as solid objects.
Instancing should now be enabled by putting a line `#pragma instanced`
in both the vertex and fragment shader.